GitHub Actions (#71)
* Add basic test to Runner

* Add runner.yaml test job

* Add Python version to setup name

* Deploy container images to Docker hub and GHCR

* Fix context

* Add QEMU

* Generate tag for a pull request

* Controller workflow

* Actually set COMP_GOOS and COMP_GOARCH

* Different split technique

* :facepalm:

* Maybe?

* Debug

* why

* Hmmm

* Bad

* Cut out the os and architecture

* Solve failed 32-bit build

* Update test

* Use temp dir as config dir

* Add QEMU back

* Try a full path for artifact upload

* Upload build artifacts to a release

* Build executables and upload to a GitHub Release

* Full path

* Remove debug ls

* Remove pull_request event

+++ controller/config/config.go
@@ -14,8 +14,8 @@ const Version string = "0.0.1-alpha.1"
 
 // Settings is used to represent how settings are saved to a file
 type Settings struct {
-	HealthCheckInterval int
-	HealthCheckTimeout  int
+	HealthCheckInterval uint64
+	HealthCheckTimeout  uint64
 	LogVerbosity        string
 	SmallerFiles        bool
 }
@@ -62,8 +62,8 @@ func SaveGlobal() error {
 // DefaultSettings returns a "constant" settings struct with sensible defaults
 func DefaultSettings() Settings {
 	return Settings{
-		HealthCheckInterval: int(1 * time.Minute),
-		HealthCheckTimeout:  int(1 * time.Hour),
+		HealthCheckInterval: uint64(1 * time.Minute),
+		HealthCheckTimeout:  uint64(1 * time.Hour),
 		LogVerbosity:        "INFO",
 		SmallerFiles:        false,
 	}
